chore: review commit hooks (#1345)
**Related Ticket:** Close #1329 

### Description of Changes
It seems that there are some changes in both Husky and Yarn in how to
make them play nicely together. Husky needs to be run once for each user
before the package begins enforcing the git hooks. When it does that,
you should see a new directory `.husky/_`. For `npm` this is done
through a prepare script. Yarn doesn't support this, so Husky recommends
a `postinstall` script.
https://typicode.github.io/husky/how-to.html#manual-setup

I've added that script, as well as added a pre-push hook to run the type
check and unit tests. That's easy to remove if contributors prefer
pushing separate commits to fix/add tests, though.