Replace root.render in useThemedControl with React Portal
From the react documentation:
If your app is fully built with React, you shouldn’t need to create any more roots, or to call root.render again.

It is uncommon to call render multiple times. Usually, your components will update state instead.

When you want to render a piece of JSX in a different part of the DOM tree that isn’t a child of your component (for example, a modal or a tooltip), use createPortal instead of createRoot.

How to guide: https://18.react.dev/reference/react-dom/createPortal\#rendering-react-components-into-non-react-dom-nodes