fix: Make the banner flag path and alt text configurable (#1447)
**Related Ticket:** https://github.com/NASA-IMPACT/veda-ui/issues/1415

Make the USWDS banner flag path and alt texts configurable. The USWDS
assets are currently resolved incorrectly because they're using absolute
paths (starting with '/') which are being resolved relative to earth.gov
instead of earth.gov/ghgcenter. As an immediate fix, this PR makes the
flag path configurable so we can adjust it based on the deployment
context.

- Path resolution could potentially affect other USWDS assets beyond the
flag and in other components. We'll be tackling this via #1429 (subpath
route handling) and #1389

- Start veda-ui locally
- Change the `flagImgSrc` path in veda-config.js to
`/img/icon-https.svg`
- Restart your server
- Verify that the flag img in the banner shows the new icon
- Change it back the original flag path
- Verify that the flag icon is shown