fix: Empty story  description close #1424 (#1438)
**Related Ticket:** #1424 

### Description of Changes
Handle the case when description is empty
Add card description to search criteria
Separate out shared data structure between StoryData and Datasetdata

### Notes & Questions About Changes

It might not be very intuitive that DatasetData requires description
while StoryData doesn't. but it takes a bigger scope of change if we
want to make description optional for DatasetData. (And we might not
want it? See the problem of empty description in this comment:
https://github.com/NASA-IMPACT/veda-ui/issues/1424#issuecomment-2634734891)


### Validation / Testing
I removed `description` from life of water story for testing:
https://deploy-preview-1438--veda-ui.netlify.app/stories?search=null