fix: mailto link in mdx not working (#1470)
**Related Ticket:** [_{GHGC #643
}_](https://github.com/US-GHG-Center/veda-config-ghg/issues/643)

### Description of Changes
Any email added to the mdx was not working properly. For instance in
GHGC news and events page, clicking on the emails would result in "page
not found" error.

![image](https://github.com/user-attachments/assets/2ca3bf98-093e-4e72-8ff5-599934cf92e5)

![image](https://github.com/user-attachments/assets/773987ef-bc3a-4e61-840f-8393cba7e20f)

Hence to fix the issue, mailto regex check is added in addition to
external link check. This will in turn use anchor tag instead of Link
for mailto links, fixing the internal routing issue.

### Notes & Questions About Changes
Instead of creating a different regex check for the mailto link, its
added to the same regex that was previously available.

**Question/Note**: I found redundant `external link check` as it's used
in both the `CustomLnk` and `getLinkProps`.

### Validation / Testing
If an email is used in mdx, it can be clicked and opened in the email
client directly.