chore: Consolidate where to handle parentDatasets (#1469)
**Related Ticket:** While working on
https://github.com/NASA-IMPACT/veda-ui/pull/1465, I realized that the
logic to find parent data sets is scattered, and it isn't easy to track
them down. I believe we put this logic on the browser side because when
we first put it, parent dataset was not necessary outside of the E&A
page. Now that we have other places that use parentDataset, I think
making it a part of the data generation process makes sense at this
point.

Page for manual test - infolayer modal on e&A page:
https://earthdata.nasa.gov/dashboard/exploration?datasets=%5B%7B%22id%22%3A%22lis-global-da-evap%22%2C%22settings%22%3A%7B%22isVisible%22%3Atrue%2C%22opacity%22%3A100%2C%22analysisMetrics%22%3A%5B%7B%22id%22%3A%22mean%22%2C%22label%22%3A%22Average%22%2C%22chartLabel%22%3A%22Average%22%2C%22themeColor%22%3A%22infographicB%22%7D%2C%7B%22id%22%3A%22std%22%2C%22label%22%3A%22St+Deviation%22%2C%22chartLabel%22%3A%22St+Deviation%22%2C%22themeColor%22%3A%22infographicD%22%7D%5D%2C%22colorMap%22%3A%22viridis%22%2C%22scale%22%3A%7B%22min%22%3A0%2C%22max%22%3A0.0001%7D%7D%7D%5D&taxonomy=%7B%7D&search=&date=2021-12-01T05%3A00%3A00.000Z

This needs to go with next-veda-ui pr :
https://github.com/NASA-IMPACT/next-veda-ui/compare/generate-parentdataset?expand=1