feat: Allow custom layers for Map Block (#1754)
**Related Ticket:** https://github.com/NASA-IMPACT/veda-ui/issues/968 😬 

### Description of Changes
While preparing the end of PI demo, I realized that we need a small
tweak on map component to allow custom layers. I don't see any harm on
it, other than maybe we should make the limitation of children component
of MapBlock clear? (it can't be jsx component, it should follow the
pattern of adding map style/)

### Notes & Questions About Changes
I am only allowing custom layers for the main map layer (not for
comparison layer) - would this be a problem?

### Validation / Testing
Check Sandbox page:
https://deploy-preview-1754--veda-ui.netlify.app/sandbox/custom-layer